requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.128.4
h11==0.16.0
idna==3.11
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
pydantic_settings==2.12.0
python-dotenv==1.2.1